+++ package/src/core/copyFiles.ts
@@ -9,15 +9,15 @@ const packageDirPath = defaults.packageDirPath
 /** Copy cms files
  * step 1: copy src/cms
  * step 2: copy src/routes/admin
- */ 
+ */
 export default async function copyFileSyncFunc(){
     const cmsPath = `${packageDirPath}/files/cms`
     const cmsPathProject = `${projectSrcPath}/cms`
     const adminPath = `${packageDirPath}/files/admin`
     const adminPathProject = `${projectSrcPath}/routes/admin`
     // Remove old folders
-    await fs.remove(cmsPath)
-    await fs.remove(cmsPathProject)
+    if(fs.existsSync(cmsPathProject)) await fs.remove(cmsPathProject)
+    if(fs.existsSync(adminPathProject)) await fs.remove(adminPathProject)
     // Copy folders
     fs.copySync(cmsPath,cmsPathProject)
     fs.copySync(adminPath,adminPathProject)
